fix(datasourceimpl) : validation to invalids id

+++ src/infrastructure/datasources/chillNews.datasource.impl.ts
@@ -1,6 +1,7 @@
 import { ChillNewsDatasource, ChillNewsDto, ChillNewsEntity, CustomErr, IChillNewsUpdateData } from "../../domain";
 import { chillNewsModel } from "../../databases/mongodb";
 import { ChillNewsMapper } from "../mappers";
+import * as mongoose from 'mongoose';
 
 export class ChillNewsDatasourceImpl implements ChillNewsDatasource {
 
@@ -20,12 +21,18 @@ export class ChillNewsDatasourceImpl implements ChillNewsDatasource {
 
     async getOne( id : string ) : Promise<ChillNewsEntity> {
         try {
-            const chillNews = await chillNewsModel.findById(id).exec();
 
+            const validation = mongoose.isValidObjectId(id);
+
+            if(!validation) throw CustomErr.badRequest('El id no es valido');
+
+            const chillNews = await chillNewsModel.findById({ _id : id}).exec();
+          
             if(!chillNews) throw CustomErr.badRequest('No se encontro el posts');
 
             return ChillNewsMapper.singleChillNewsMapper( chillNews );
         } catch (error) {
+
             if(error instanceof CustomErr){
                 throw error;
             }
@@ -34,13 +41,22 @@ export class ChillNewsDatasourceImpl implements ChillNewsDatasource {
         }
     }
     async update( payload: IChillNewsUpdateData): Promise<ChillNewsEntity> {
+
         try {
-            const updatedChillNews = await chillNewsModel.updateOne({ _id : payload.id }, payload.ChillNewsDto );
+
+            const validation = mongoose.isValidObjectId(payload.id);
+
+            if(!validation) throw CustomErr.badRequest('El id no es valido');
+
+            const updatedChillNews = await chillNewsModel.findOneAndUpdate({ _id : payload.id }, payload.ChillNewsDto );
 
             if(!updatedChillNews) throw CustomErr.badRequest('No se encontre la chillNews');
 
+            updatedChillNews.save();
+
             return ChillNewsMapper.singleChillNewsMapper( updatedChillNews as any );
         } catch (error) {
+
             if(error instanceof CustomErr){
                 throw error;
             }
@@ -52,6 +68,11 @@ export class ChillNewsDatasourceImpl implements ChillNewsDatasource {
     
     async remove( id : string ): Promise<void> {
         try {
+
+            const validation = mongoose.isValidObjectId(id);
+
+            if(!validation) throw CustomErr.badRequest('El id no es valido');
+
             await chillNewsModel.deleteOne({ _id : id});
 
             return 
@@ -65,13 +86,13 @@ export class ChillNewsDatasourceImpl implements ChillNewsDatasource {
     }
 
     async add ( ChillNewsDto: ChillNewsDto ): Promise<any> {
-        const { userId, avatar, chillNewContent } = ChillNewsDto;
+        const { userId, avatar, content } = ChillNewsDto;
 
        try {
          const newChillNews = await chillNewsModel.create({
              userId : userId,
              avatar : avatar,
-             content : chillNewContent
+             content : content
          })
  
          await newChillNews.save();
